requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
blas=1.0=mkl
bottleneck=1.3.7=py312ha883a20_0
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.6.2=hbcca054_0
certifi=2024.6.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.2.1=py312h8572e83_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cudnn=8.9.7.29=h092f7fd_3
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fontconfig=2.14.2=h14ed4e7_0
fonttools=4.53.0=py312h9a8786e_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.6.0=pyhff2d567_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.1=h6a678d5_0
idna=3.7=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py312h06a4308_0
jpeg=9e=h5eee18b_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py312h8572e83_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h59595ed_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=13.2.0=h77fa898_10
libgfortran-ng=13.2.0=h69a702a_10
libgfortran5=13.2.0=h3d2ce59_10
libglib=2.78.4=hdc74915_0
libgomp=13.2.0=h77fa898_10
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hdb19cb5_3
libmagma=2.7.2=h173bb3b_2
libmagma_sparse=2.7.2=h173bb3b_3
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.27=pthreads_h413a1c8_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=h08a7969_0
libsqlite=3.46.0=hde9e2c9_0
libstdcxx-ng=13.2.0=hc0a3c3a_10
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=hd590300_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hfdd30dd_2
libxslt=1.1.37=h5eee18b_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=14.0.6=h9e868ea_0
lxml=5.2.1=py312hdbbb534_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.8.4=py312h06a4308_0
matplotlib-base=3.8.4=py312h20ab3a6_2
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.20=hf484d3e_1001
nccl=2.21.5.1=h3a97aeb_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py312h06a4308_0
numexpr=2.8.7=py312hf827012_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.1=h4ab18f5_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py312h526ad5a_0
pcre2=10.42=hebb0a14_1
pillow=10.3.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
ply=3.11=py312h06a4308_1
pyparsing=3.1.2=pyhd8ed1ab_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=4_cp312
pytorch=2.3.1=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.1=py312h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.32.2=py312h06a4308_0
setuptools=69.5.1=py312h06a4308_0
sip=6.7.12=py312h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
torchaudio=2.3.1=py312_cu121
torchvision=0.18.1=py312_cu121
tornado=6.3.3=py312h5eee18b_0
tqdm=4.66.4=py312he106c6f_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=py312h06a4308_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.6=ha6fb4c9_0
